# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * REST API for a Telegram bot control panel
 * OpenAPI spec version: 0.1.0
 */
import type { ProductStatus } from './productStatus';

export interface Product {
  id: number;
  title: string;
  description: string;
  category: string;
  price: number;
  currency: string;
  /** @nullable */
  imageUrl: string | null;
  stock: number;
  status: ProductStatus;
  createdAt: Date;
}