# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * REST API for a Telegram bot control panel
 * OpenAPI spec version: 0.1.0
 */
import type { ProductStatus } from './productStatus';

export interface ProductInput {
  /** @minLength 1 */
  title: string;
  description: string;
  /** @minLength 1 */
  category: string;
  /** @minimum 0 */
  price: number;
  /**
     * @minLength 3
     * @maxLength 3
     */
  currency: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @minimum 0 */
  stock: number;
  status: ProductStatus;
}