# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * REST API for a Telegram bot control panel
 * OpenAPI spec version: 0.1.0
 */
import type { UserStatus } from './userStatus';

export interface UserUpdate {
  /** @minLength 1 */
  name?: string;
  /** @nullable */
  username?: string | null;
  status?: UserStatus;
}